syntax highlighting

+++ install.sh
@@ -13,7 +13,12 @@ if [ -e ~/.vim/after/ftplugin ]; then
     echo "error: after/ftplugin exists"
     exit
 fi
+if [ -e ~/.vim/after/ftdetect ]; then
+    echo "error: after/ftplugin exists"
+    exit
+fi
 ln -s $(pwd)/ftplugin ~/.vim/after
+ln -s $(pwd)/ftdetect ~/.vim/after
 ln -s $(pwd)/mybashrc ~/.mybashrc
 if [ ! -d ~/.vim/bundle/Vundle.vim ]; then
     git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
